working on human fastq linking

+++ app/models/study_file.rb
@@ -4,10 +4,13 @@ class StudyFile
   include Mongoid::Paperclip
   include Rails.application.routes.url_helpers
 
+  # constants, used for statuses and file types
   STUDY_FILE_TYPES = ['Cluster Coordinates', 'Cluster Assignments', 'Expression Matrix', 'Marker Gene List', 'Fastq', 'Documentation', 'Other']
   PARSEABLE_TYPES = ['Cluster Coordinates', 'Expression Matrix', 'Marker Gene List']
   UPLOAD_STATUSES = %w(new uploading uploaded)
+  PARSE_STATUSES = %w(unparsed parsing parsed)
 
+  # associations
   belongs_to :study, index: true
   has_many :clusters, dependent: :destroy
   has_many :single_cells, dependent: :destroy
@@ -16,14 +19,18 @@ class StudyFile
   has_many :precomputed_scores, dependent: :destroy
   has_many :temp_file_downloads
 
+  # field definitions
   field :name, type: String
   field :path, type: String
   field :description, type: String
   field :file_type, type: String
   field :url_safe_name, type: String
   field :status, type: String
-  field :parsed, type: Boolean, default: false
+  field :parse_status, type: String, default: 'unparsed'
+  field :human_fastq_url, type: String
+  field :human_data, type: Boolean, default: false
 
+  # callbacks
   before_create   :make_data_dir
   before_create   :set_file_name_and_url_safe_name
   after_save      :check_public?
@@ -43,14 +50,20 @@ class StudyFile
     attachment.instance.url_safe_name
   end
 
+  # return public url if study is public, otherwise redirect to create templink download url
   def download_path
-    if self.study.public?
-      self.upload.url
+    if self.upload_file_name.nil?
+      self.human_fastq_url
     else
-      download_private_file_path(self.study.url_safe_name, self.upload_file_name)
+      if self.study.public?
+        self.upload.url
+      else
+        download_private_file_path(self.study.url_safe_name, self.upload_file_name)
+      end
     end
   end
 
+  # JSON response for jQuery uploader
   def to_jq_upload
     {
         '_id' => self._id,
@@ -66,6 +79,10 @@ class StudyFile
     PARSEABLE_TYPES.include?(self.file_type)
   end
 
+  def parsed?
+    self.parse_status == 'parsed'
+  end
+
   private
 
   def make_data_dir
@@ -75,22 +92,31 @@ class StudyFile
     end
   end
 
+  # set filename and construct url safe name from study
   def set_file_name_and_url_safe_name
-    self.name = self.upload_file_name
+    if self.upload_file_name.nil?
+      self.status = 'uploaded'
+    else
+      self.name = self.upload_file_name
+    end
     self.url_safe_name = self.study.url_safe_name
   end
 
   # add symlink if study is public
   def check_public?
-    if self.study.public? && self.status == 'uploaded'
-      FileUtils.ln_sf(self.upload.path, File.join(self.study.data_public_path, self.upload_file_name))
+    unless self.upload_file_name.nil?
+      if self.study.public? && self.status == 'uploaded'
+        FileUtils.ln_sf(self.upload.path, File.join(self.study.data_public_path, self.upload_file_name))
+      end
     end
   end
 
   # clean up any symlinks before deleting a study file
   def remove_public_symlink
-    if self.study.public?
-      FileUtils.rm_f(File.join(self.study.data_public_path, self.upload_file_name))
+    unless self.upload_file_name.nil?
+      if self.study.public?
+        FileUtils.rm_f(File.join(self.study.data_public_path, self.upload_file_name))
+      end
     end
   end
